requirements.txt
aiohttp==0.16.5
aiohttp-jinja2==0.4.3
chardet==2.3.0
cssselect==0.9.1
Jinja2==2.7.3
lxml==3.4.4
MarkupSafe==0.23
PyExecJS==1.1.0
pyquery==1.2.9
PyReact==0.5.2
requests==2.7.0
